backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2 
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.0